refactor: move dataset quality prompt modules to the scorers.dataset_quality.prompts directory
